requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
boto3==1.40.18
botocore==1.40.18
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
defusedxml==0.7.1
Django==5.2.5
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
frozenlist==1.7.0
google-auth==2.40.3
idna==3.10
inflection==0.5.1
jmespath==1.0.1
kombu==5.5.4
multidict==6.6.4
mysql-connector-python==9.4.0
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
twilio==9.7.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.1
